+++ Servers/Filters/vtkPVTrackballRotate.cxx
@@ -71,8 +71,8 @@ void vtkPVTrackballRotate::OnMouseMove(int x, int y, vtkRenderer *ren,
   transform->Identity();
   transform->Translate(this->Center[0], this->Center[1], this->Center[2]);
   
-  float dx = rwi->GetLastEventPosition()[0] - x;
-  float dy = rwi->GetLastEventPosition()[1] - y;
+  int dx = rwi->GetLastEventPosition()[0] - x;
+  int dy = rwi->GetLastEventPosition()[1] - y;
   
   // azimuth
   camera->OrthogonalizeViewUp();
